fix: add useCreateInk
